requirements.txt
numpy~=1.20.3
pep517
setuptools~=56.0.0
Sphinx
recommonmark
sphinx_rtd_theme
twine
requests~=2.25.1
websocket-client
